requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.6.1
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.5.2
joblib==1.4.2
python-dotenv==1.0.1
xgboost==2.0.3
passlib[bcrypt]
python-jose[cryptography]
python-multipart
email-validator==2.2.0
